paperreview-env/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
httpx>=0.25.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
